+++ NEWS.md
@@ -1,6 +1,6 @@
 # colorpath 0.0.0 (development version)
 
-* Added a plot method for a `surface_hl()` function, uses `colorspace::hclplot()`. (#31)
+* Added functions `data_surface_hl()`, `plot_surface_hl()` that work with hue-luminance surface functions. (#31)
 
 * Added functions `surface_hl() ` and `df_hcl()` to facilitate creating sets of consistent multi-hue palettes. (#29)